+++ .php-cs-fixer.dist.php
@@ -11,11 +11,15 @@ return (new PhpCsFixer\Config)
         '@Symfony' => true,
         'array_syntax' => ['syntax' => 'short'],
         'ordered_class_elements' => true,
+        'ordered_imports' => ['imports_order' => ['class', 'function', 'const']],
         'yoda_style' => false,
         'native_function_invocation' => [
             'include' => ['@compiler_optimized']
         ],
         'php_unit_method_casing' => false,
-        'declare_strict_types' => true
+        'declare_strict_types' => true,
+        'global_namespace_import' => [
+            'import_classes' => false,
+        ],
     ])
 ;