+++ config/services.php
@@ -17,11 +17,12 @@ use Baldinof\RoadRunnerBundle\Integration\Symfony\StreamedResponseListener;
 use Baldinof\RoadRunnerBundle\Reboot\KernelRebootStrategyInterface;
 use Baldinof\RoadRunnerBundle\RoadRunnerBridge\HttpFoundationWorker;
 use Baldinof\RoadRunnerBundle\RoadRunnerBridge\HttpFoundationWorkerInterface;
-use Baldinof\RoadRunnerBundle\Worker\Dependencies;
-use Baldinof\RoadRunnerBundle\Worker\GrpcWorker;
+use Baldinof\RoadRunnerBundle\Worker\GrpcWorker as InternalGrpcWorker;
 use Baldinof\RoadRunnerBundle\Worker\GrpcWorkerInterface;
-use Baldinof\RoadRunnerBundle\Worker\Worker;
-use Baldinof\RoadRunnerBundle\Worker\WorkerInterface;
+use Baldinof\RoadRunnerBundle\Worker\HttpDependencies;
+use Baldinof\RoadRunnerBundle\Worker\HttpWorker as InternalHttpWorker;
+use Baldinof\RoadRunnerBundle\Worker\WorkerRegistry;
+use Baldinof\RoadRunnerBundle\Worker\WorkerRegistryInterface;
 use Psr\Log\LoggerInterface;
 use Sentry\SentryBundle\EventListener\TracingRequestListener;
 use Sentry\State\HubInterface;
@@ -37,6 +38,7 @@ use Spiral\RoadRunner\Metrics\Metrics;
 use Spiral\RoadRunner\Metrics\MetricsInterface;
 use Spiral\RoadRunner\Worker as RoadRunnerWorker;
 use Spiral\RoadRunner\WorkerInterface as RoadRunnerWorkerInterface;
+use Symfony\Component\HttpKernel\Kernel;
 use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 
 // Polyfill of the `service()` function introduced in Symfony 5.1 when using older version
@@ -75,7 +77,10 @@ return static function (ContainerConfigurator $container) {
     $services->set(HttpFoundationWorkerInterface::class, HttpFoundationWorker::class)
         ->args([service(HttpWorkerInterface::class)]);
 
-    $services->set(WorkerInterface::class, Worker::class)
+    $services->set(WorkerRegistryInterface::class, WorkerRegistry::class)
+        ->public();
+
+    $services->set(InternalHttpWorker::class)
         ->public() // Manually retrieved on the DIC in the Worker if the kernel has been rebooted
         ->tag('monolog.logger', ['channel' => BaldinofRoadRunnerExtension::MONOLOG_CHANNEL])
         ->args([
@@ -84,7 +89,14 @@ return static function (ContainerConfigurator $container) {
             service(HttpFoundationWorkerInterface::class),
         ]);
 
-    $services->set(Dependencies::class)
+    $services
+        ->get(WorkerRegistryInterface::class)
+        ->call('registerWorker', [
+            Environment\Mode::MODE_HTTP,
+            service(InternalHttpWorker::class),
+        ]);
+
+    $services->set(HttpDependencies::class)
         ->public() // Manually retrieved on the DIC in the Worker if the kernel has been rebooted
         ->args([
             service(MiddlewareStack::class),
@@ -93,7 +105,7 @@ return static function (ContainerConfigurator $container) {
         ]);
 
     $services->set(WorkerCommand::class)
-        ->args([service(WorkerInterface::class)])
+        ->args([service(InternalHttpWorker::class)])
         ->autoconfigure();
 
     $services->set(KernelHandler::class)
@@ -107,13 +119,6 @@ return static function (ContainerConfigurator $container) {
     $services->alias(RequestHandlerInterface::class, MiddlewareStack::class);
 
     $services->set(NativeSessionMiddleware::class);
-
-    $services->set(StreamedResponseListener::class)
-        ->decorate('streamed_response_listener')
-        ->args([
-            service(StreamedResponseListener::class.'.inner'),
-            '%env(default::RR_MODE)%',
-        ]);
     
     $services->set(SentryTracingRequestListener::class)
         ->decorate(TracingRequestListener::class, null, 0, ContainerInterface::IGNORE_ON_INVALID_REFERENCE)
@@ -121,6 +126,16 @@ return static function (ContainerConfigurator $container) {
             service('.inner'),
             service(HubInterface::class),
         ]);
+        
+    // @phpstan-ignore-next-line - PHPStan says this is always true, but the constant value depends on the currently installed Symfony version
+    if (Kernel::VERSION_ID < 60100) {
+        $services->set(StreamedResponseListener::class)
+            ->decorate('streamed_response_listener')
+            ->args([
+                service(StreamedResponseListener::class.'.inner'),
+                '%env(default::RR_MODE)%',
+            ]);
+    }
 
     if (interface_exists(GrpcServiceInterface::class)) {
         $services->set(GrpcServiceProvider::class);
@@ -131,7 +146,7 @@ return static function (ContainerConfigurator $container) {
                 service(GrpcInvoker::class),
             ]);
 
-        $services->set(GrpcWorkerInterface::class, GrpcWorker::class)
+        $services->set(GrpcWorkerInterface::class, InternalGrpcWorker::class)
             ->public() // Manually retrieved on the DIC in the Worker if the kernel has been rebooted
             ->tag('monolog.logger', ['channel' => BaldinofRoadRunnerExtension::MONOLOG_CHANNEL])
             ->args([
@@ -139,6 +154,13 @@ return static function (ContainerConfigurator $container) {
                 service(RoadRunnerWorkerInterface::class),
                 service(GrpcServiceProvider::class),
                 service(GrpcServer::class),
+            ]);
+
+        $services
+            ->get(WorkerRegistryInterface::class)
+            ->call('registerWorker', [
+                Environment\Mode::MODE_GRPC,
+                service(GrpcWorkerInterface::class),
             ]);
     }
 };