+++ src/DependencyInjection/Configuration.php
@@ -15,6 +15,7 @@ class Configuration implements ConfigurationInterface
 
     public const KERNEL_REBOOT_STRATEGY_ALWAYS = 'always';
     public const KERNEL_REBOOT_STRATEGY_ON_EXCEPTION = 'on_exception';
+    public const KERNEL_REBOOT_STRATEGY_MAX_JOBS = 'max_jobs';
 
     public function getConfigTreeBuilder(): TreeBuilder
     {
@@ -27,19 +28,30 @@ class Configuration implements ConfigurationInterface
                 ->arrayNode('kernel_reboot')
                     ->addDefaultsIfNotSet()
                     ->children()
-                        ->scalarNode('strategy')
+                        ->arrayNode('strategy')
                             ->info(sprintf(
-                                'Possible values are "%s", "%s" or any service that implements "%s"/',
+                                'Possible values are "%s", "%s", "%s" or any service that implements "%s"/',
                                 self::KERNEL_REBOOT_STRATEGY_ALWAYS,
                                 self::KERNEL_REBOOT_STRATEGY_ON_EXCEPTION,
+                                self::KERNEL_REBOOT_STRATEGY_MAX_JOBS,
                                 KernelRebootStrategyInterface::class
                             ))
-                            ->defaultValue(self::KERNEL_REBOOT_STRATEGY_ON_EXCEPTION)
+                            ->defaultValue([self::KERNEL_REBOOT_STRATEGY_ON_EXCEPTION])
+                            ->beforeNormalization()->castToArray()->end()
+                            ->scalarPrototype()->end()
                         ->end()
                         ->arrayNode('allowed_exceptions')
                             ->info('Only used when `reboot_kernel.strategy: on_exception`. Exceptions defined here will not cause kernel reboots.')
                             ->defaultValue([])
                             ->scalarPrototype()->end()
+                        ->end()
+                        ->scalarNode('max_jobs')
+                            ->info('Only used when `reboot_kernel.strategy: max_jobs`. Maximum numbers of jobs before kernel reboot')
+                            ->defaultValue(1000)
+                        ->end()
+                        ->scalarNode('max_jobs_dispersion')
+                            ->info('Only used when `reboot_kernel.strategy: max_jobs`. Dispersion persent')
+                            ->defaultValue(0.2)
                         ->end()
                     ->end()
                 ->end()