+++ src/Worker/GrpcWorkerInterface.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace Baldinof\RoadRunnerBundle\Worker;
 
-interface GrpcWorkerInterface
+/**
+ * @deprecated and will be removed at 3.0, use Baldinof\RoadRunnerBundle\Worker\WorkerInterface instead
+ */
+interface GrpcWorkerInterface extends WorkerInterface
 {
-    public function start(): void;
 }