+++ tests/Http/KernelHandlerTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Tests\Baldinof\RoadRunnerBundle\Http;
 
-use function Baldinof\RoadRunnerBundle\consumes;
 use Baldinof\RoadRunnerBundle\Http\KernelHandler;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\HttpFoundation\Request;
@@ -12,6 +11,8 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\HttpKernelInterface;
 use Symfony\Component\HttpKernel\TerminableInterface;
 use Tests\Baldinof\RoadRunnerBundle\Utils\CallableHttpKernel;
+
+use function Baldinof\RoadRunnerBundle\consumes;
 
 class KernelHandlerTest extends TestCase
 {