+++ tests/Integration/Sentry/SentryMiddlewareTest.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace Tests\Baldinof\RoadRunnerBundle\Integration\Sentry;
 
-use function Baldinof\RoadRunnerBundle\consumes;
 use Baldinof\RoadRunnerBundle\Integration\Sentry\SentryMiddleware;
-use Closure;
 use GuzzleHttp\Promise\Promise;
 use GuzzleHttp\Promise\PromiseInterface;
 use PHPUnit\Framework\TestCase;
@@ -18,18 +16,19 @@ use Sentry\SentrySdk;
 use Sentry\State\Hub;
 use Sentry\Transport\TransportFactoryInterface;
 use Sentry\Transport\TransportInterface;
-use SplStack;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\HttpKernelInterface;
 use Tests\Baldinof\RoadRunnerBundle\Utils\CallableHttpKernel;
 
+use function Baldinof\RoadRunnerBundle\consumes;
+
 final class SentryMiddlewareTest extends TestCase
 {
-    public static SplStack $collectedEvents;
+    public static \SplStack $collectedEvents;
 
-    /** @var Closure(): Response */
-    public Closure $onRequest;
+    /** @var \Closure(): Response */
+    public \Closure $onRequest;
 
     private HttpKernelInterface $handler;